Switch graphic & key management to OpenGL/GLFW3, correct sprite behaviour, still not perfect, but way better (no more issue with overlapping and overflow is correctly handled now)  ->PPU still need a complete rewrite.  Bump version to 0.7!

+++ src/include/ppu/ppu.h
@@ -5,7 +5,7 @@
  *  Define and emulate the PPU (Picture Processing Unit) of the real NES
  * 
  *  Created by Manoel TRAPIER.
- *  Copyright (c) 2003-2008 986Corp. All rights reserved.
+ *  Copyright (c) 2003-2016 986-Studio. All rights reserved.
  *
  *  $LastChangedDate$
  *  $Author$
@@ -21,10 +21,10 @@
 
 typedef struct PPU_Sprite_
 {
-    byte y;
-    byte tileid;
-    byte flags;
-    byte x;
+    uint8_t y;
+    uint8_t tileid;
+    uint8_t flags;
+    uint8_t x;
 } PPU_Sprite;
 
 /*
@@ -32,13 +32,13 @@ PPU must be initialized after memory initialisation..
 */
 int ppu_init();
 
-int ppu_hblank(int scanline);
+int ppu_hblank(uint16_t scanline);
 
-byte ppu_readReg(byte id);
+uint8_t ppu_readReg(uint8_t id);
 
-void ppu_writeReg(byte id, byte val);
+void ppu_writeReg(uint8_t id, uint8_t val);
 
-void ppu_fillSprRamDMA(byte value);
+void ppu_fillSprRamDMA(uint8_t value);
 
 #define PPU_MIRROR_HORIZTAL 0
 #define PPU_MIRROR_VERTICAL 1
@@ -52,15 +52,15 @@ void ppu_fillSprRamDMA(byte value);
 #define PPU_SCMODE_NORMAL 1
 #define PPU_SCMODE_FOURSC 2
 
-void ppu_setMirroring(byte direction);
-void ppu_setSingleScreen(byte screen);
-void ppu_setScreenMode(byte mode);
+void ppu_setMirroring(uint8_t direction);
+void ppu_setSingleScreen(uint8_t screen);
+void ppu_setScreenMode(uint8_t mode);
 
 
-PPU_Sprite ppu_getSprite(unsigned short i);
+PPU_Sprite ppu_getSprite(uint16_t i);
 
-unsigned char ppu_memoryRead(byte page, byte addr);
-void          ppu_memoryWrite(byte page, byte addr, byte value);
+unsigned char ppu_memoryRead(uint8_t page, uint8_t addr);
+void          ppu_memoryWrite(uint8_t page, uint8_t addr, uint8_t value);
 
 void ppu_debugSprites();
 void ppu_debugColor();
